Use uint16 in block meta

+++ tempodb/encoding/streaming_block.go
@@ -132,7 +132,7 @@ func (c *StreamingBlock) Complete(ctx context.Context, tracker backend.AppendTra
 
 	meta.TotalRecords = uint32(len(records)) // casting
 	meta.IndexPageSize = uint32(c.cfg.IndexPageSizeBytes)
-	meta.BloomShardCount = uint8(c.bloom.GetShardCount())
+	meta.BloomShardCount = uint16(c.bloom.GetShardCount())
 
 	return bytesFlushed, writeBlockMeta(ctx, w, meta, indexBytes, c.bloom)
 }
